CELIX-370: disable -Wall -Werror for C++. apparently cpputest still contains warnings..

+++ utils/private/test/celix_threads_test.cpp
@@ -67,7 +67,7 @@ static int celix_thread_t_equals(const void * object, const void * compareTo){
 static char * celix_thread_t_toString(const void * object){
 	celix_thread_t * thread = (celix_thread_t*) object;
 	char buff[512];
-	snprintf(buff, 512, "thread: %lu, threadInitialized: %s", thread->thread, (thread->threadInitialized ? "true" : "false"));
+	snprintf(buff, 512, "thread: %lu, threadInitialized: %s", (unsigned long)thread->thread, (thread->threadInitialized ? "true" : "false"));
 
 	return my_strdup(buff);
 }
